# backend-parry-symfony/config/openapi/game-api.yaml
# config/openapi/game-api.yaml
openapi: 3.0.0
info:
  title: PARRY Game API
  description: API pour la gestion des parties et des IA du jeu PARRY
  version: 1.0.0

tags:
  - name: Game
    description: Gestion des parties et interactions IA
  - name: Auth
    description: Authentification et gestion des utilisateurs

paths:
  /api/game/round/question:
    post:
      tags:
        - Game
      summary: Générer une question pour une manche
      description: L'IA génère une question originale et engageante pour démarrer une manche
      operationId: generateQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                round_number:
                  type: integer
                  example: 1
                  description: Numéro de la manche actuelle
                previous_questions:
                  type: array
                  items:
                    type: string
                  example:
                    - "Quel est ton film préféré ?"
                    - "Quelle est ta destination de rêve ?"
                  description: Liste des questions déjà posées (pour éviter les doublons)
      responses:
        '200':
          description: Question générée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  question:
                    type: string
                    example: "Quel est ton plat préféré ?"
              examples:
                example1:
                  summary: Question sur la nourriture
                  value:
                    question: "Si tu devais manger un seul plat pour le reste de ta vie, ce serait quoi ?"
                example2:
                  summary: Question sur les loisirs
                  value:
                    question: "Quelle est la dernière série que tu as binge-watchée ?"
        '500':
          description: Erreur lors de la génération
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/game/round/ai-response:
    post:
      tags:
        - Game
      summary: L'IA génère une réponse à une question
      description: L'IA répond de manière convaincante pour passer pour un humain
      operationId: generateAIResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  example: "Quel est ton film préféré ?"
      responses:
        '200':
          description: Réponse générée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    example: "Bah euh... j'adore Inception, c'est trop bien 🎬"
              examples:
                casual:
                  summary: Réponse décontractée
                  value:
                    response: "Ah bah moi c'est Interstellar, genre j'ai pleuré à la fin 😭"
                neutral:
                  summary: Réponse neutre
                  value:
                    response: "J'hésite entre plusieurs mais je dirais The Dark Knight"
        '400':
          description: Question manquante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/game/moderate:
    post:
      tags:
        - Game
      summary: Modérer un contenu
      description: Analyse un contenu (question ou réponse) pour détecter les contenus inappropriés
      operationId: moderateContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "Mon contenu à modérer"
                type:
                  type: string
                  enum:
                    - question
                    - response
                  example: response
                  default: response
      responses:
        '200':
          description: Résultat de la modération
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_safe:
                    type: boolean
                    example: true
                  flagged_reasons:
                    type: array
                    items:
                      type: string
                    example: []
                  severity:
                    type: string
                    enum:
                      - none
                      - low
                      - medium
                      - high
                    example: none
              examples:
                safe:
                  summary: Contenu sûr
                  value:
                    is_safe: true
                    flagged_reasons: []
                    severity: none
                unsafe:
                  summary: Contenu inapproprié
                  value:
                    is_safe: false
                    flagged_reasons:
                      - profanity
                      - hate_speech
                    severity: high

  /api/game/round/ai-vote:
    post:
      tags:
        - Game
      summary: L'IA vote pour un joueur suspect
      description: L'IA analyse les réponses et vote stratégiquement pour un joueur (sans se faire démasquer)
      operationId: generateAIVote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
                - players
              properties:
                question:
                  type: string
                  example: "Quel est ton film préféré ?"
                players:
                  type: array
                  items:
                    type: object
                    properties:
                      player_id:
                        type: string
                        example: player_abc123
                      response:
                        type: string
                        example: "J'aime beaucoup les films de science-fiction"
                      is_ai:
                        type: boolean
                        example: false
                  example:
                    - player_id: player_abc123
                      response: "J'aime beaucoup les films de science-fiction"
                      is_ai: false
                    - player_id: player_def456
                      response: "Mon film préféré est Inception car il explore brillamment..."
                      is_ai: false
      responses:
        '200':
          description: Vote généré
          content:
            application/json:
              schema:
                type: object
                properties:
                  voted_for:
                    type: string
                    example: player_def456
                    description: ID du joueur pour qui l'IA vote
        '400':
          description: Données manquantes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Message d'erreur"

  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authentification par token JWT.
        Format: `Authorization: Bearer <token>`

security:
  - Bearer: []